requirements.txt
astroid==2.0.4
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
coveralls==1.7.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
djangorestframework==3.9.4
docopt==0.6.2
dodgy==0.1.9
entrypoints==0.3
ez-setup==0.9
facebook-sdk==3.1.0
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
pep8-naming==0.4.1
prospector==1.1.6.2
psycopg2==2.7.7
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.1.1
pylint-celery==0.3
pylint-django==2.0.2
pylint-flask==0.5
pylint-plugin-utils==0.5
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
requirements-detector==0.6
setoptconf==0.2.0
six==1.12.0
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.2
whitenoise==4.1.2
wrapt==1.11.1
